+++ bin/build_languages.dart
@@ -4,8 +4,15 @@ import 'package:recase/recase.dart';
 
 import 'package:language_picker/languages.dart';
 
+toGetterName(String languageName) => languageName
+    // Remove commas and parentheses
+    .replaceAll(RegExp('[,\(\)]'), '')
+    // Replace punctuated name
+    .replaceAll('ü', 'u')
+    .camelCase;
+
 /// Reads languages.json and generates lib/languages.g.dart, which contains
-/// Languages.countryName constants.
+/// Languages.countryName constants and the Languages.defaultLanguages list.
 main() {
   final contents = File('lib/languages.json').readAsStringSync();
   final jsonLanguages = jsonDecode(contents);
@@ -13,20 +20,23 @@ main() {
       .map<Language>((m) => Language.fromMap(Map<String, String>.from(m)))
       .toList();
   final getters = languages.map((l) {
-    final escapedName = l.name
-        // Remove commas and parentheses
-        .replaceAll(RegExp('[,\(\)]'), '')
-        // Replace punctuated name
-        .replaceAll('ü', 'u')
-        .camelCase;
-    return "  static Language get $escapedName => Language('${l.isoCode}', '${l.name}');";
+    final getterName = toGetterName(l.name);
+    return "  static Language get $getterName => Language('${l.isoCode}', '${l.name}');";
   });
-  final extension = '''
+
+  final defaultLanguages = '''
+static List<Language> defaultLanguages =
+  [${languages.map((l) => 'Languages.' + toGetterName(l.name)).join(',\n')}];
+''';
+
+  final staticClass = '''
 import 'languages.dart';
 
 class Languages {
 ${getters.join('\n')}
+
+$defaultLanguages
 }
 ''';
-  File('lib/languages.g.dart').writeAsStringSync(extension);
+  File('lib/languages.g.dart').writeAsStringSync(staticClass);
 }