+++ lib/language_picker_cupertino.dart
@@ -60,7 +60,7 @@ class LanguagePickerCupertino extends StatefulWidget {
   final FixedExtentScrollController? scrollController;
 
   /// List of languages available in this picker.
-  final List<Map<String, String>>? languagesList;
+  final List<Language>? languages;
 
   const LanguagePickerCupertino({
     Key? key,
@@ -75,7 +75,7 @@ class LanguagePickerCupertino extends StatefulWidget {
     this.useMagnifier,
     this.magnification,
     this.scrollController,
-    this.languagesList,
+    this.languages,
   }) : super(key: key);
 
   @override
@@ -89,8 +89,7 @@ class _CupertinoLanguagePickerState extends State<LanguagePickerCupertino> {
   @override
   void initState() {
     super.initState();
-    final languageList = widget.languagesList ?? defaultLanguagesList;
-    _allLanguages = languageList.map((item) => Language.fromMap(item)).toList();
+    _allLanguages = widget.languages ?? Languages.defaultLanguages;
   }
 
   @override